upgrade to fmt 10 and patch for c++20 (but not upgrade doe to slang)

+++ lgraph/lgedge.cpp
@@ -209,11 +209,11 @@ void Node_internal::xxx(const Edge_raw *edge_raw) {
 
 // LCOV_EXCL_START
 void Node_internal::dump() const {
-  fmt::print("nid:{} pid:{} state:{} inp_pos:{} out_pos:{} root:{}\n", nid, dst_pid, state, inp_pos, out_pos, root);
+  fmt::print("nid:{} pid:{} state:{} inp_pos:{} out_pos:{} root:{}\n", (int)nid, (int)dst_pid, (int)state, (int)inp_pos, out_pos, root);
 
   const Edge_raw *out = get_output_begin();
   while (out != get_output_end()) {
-    fmt::print("  out idx:{} pid:{}\n", out->get_idx(), out->get_inp_pid());
+    fmt::print("  out idx:{} pid:{}\n", (int)out->get_idx(), (int)out->get_inp_pid());
     if (out->is_snode()) {
       out++;
     } else {
@@ -223,7 +223,7 @@ void Node_internal::dump() const {
 
   out = get_input_begin();
   while (out != get_input_end()) {
-    fmt::print("  inp idx:{} pid:{}\n", out->get_idx(), out->get_inp_pid());
+    fmt::print("  inp idx:{} pid:{}\n", (int)out->get_idx(), (int)out->get_inp_pid());
     if (out->is_snode()) {
       out++;
     } else {
